+++ driver-core/src/main/java/com/datastax/driver/core/policies/DCAwareRoundRobinPolicy.java
@@ -51,7 +51,7 @@ import com.datastax.driver.core.Statement;
  * incurs a slight overhead so the {@code LoadBalancingPolicy.RoundRobin}
  * policy could be preferred to this policy in that case.
  */
-public class DCAwareRoundRobinPolicy implements LoadBalancingPolicy {
+public class DCAwareRoundRobinPolicy implements LoadBalancingPolicy, CloseableLoadBalancingPolicy {
 
     private static final Logger logger = LoggerFactory.getLogger(DCAwareRoundRobinPolicy.class);
 
@@ -414,5 +414,10 @@ public class DCAwareRoundRobinPolicy implements LoadBalancingPolicy {
     @Override
     public void onRemove(Host host) {
         onDown(host);
+    }
+
+    @Override
+    public void close() {
+        // nothing to do
     }
 }